pyproject.toml
[tool.poetry]
name = "projectname"
version = "0.1.0"
description = "Project description"
authors = ["Martim Lobao <m@rtimlobao.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.3.0"
coverage = "^7.2.7"
docformatter = "^1.7.5"
ipython = "^8.14.0"
isort = { extras = ["colors"], version = "^5.12.0" }
mypy = "^1.4.0"
pylint = "^3.0.2"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pyupgrade = "^3.9.0"
ruff = "^0"
types-beautifulsoup4 = "^4.12.0.5"
types-pytz = "^2023.3.0.0"
types-requests = "^2.31.0.1"
yamllint = "^1.32.0"

[tool.black]
line_length = 99
preview = true

[tool.isort]
profile = "black"
color_output = true
known_first_party = ["projectname"]
line_length = 99

[tool.pyupgrade]
py311plus = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
